adding the pedometer delete feature and fixing some of the test scripts since the daily count may not be what we think it is
when the clear data command is sent, it looks like the hourly data pointer is correctly reset but the daily data is not.
In addition, when iterating through to the end of the daily data count after a delete seems to provide bad data at some
point in the past.  I will continue more testing to see if i can figure out the behavior.